added weapons service and integration with characters (1 to 1 relationship)

+++ Services/CharacterService/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
@@ -15,21 +16,30 @@ namespace dotnet_rpg.Services.CharacterService
         //add a static mock character
         public readonly IMapper _mapper;
         public readonly DataContext _context;
+        public readonly IHttpContextAccessor _httpContextAccessor;
         
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
             _mapper = mapper; 
+            _httpContextAccessor = httpContextAccessor;
         }
+        
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User
+            .FindFirstValue(ClaimTypes.NameIdentifier));
         
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
           
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+            
             _context.Characters.Add(character);
             await _context.SaveChangesAsync(); 
-            serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
             return serviceResponse;
         }
 
@@ -39,10 +49,21 @@ namespace dotnet_rpg.Services.CharacterService
             
             try
             { 
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if(character != null)
+                { 
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else { 
+                    response.Success = false; 
+                    response.Message = "Character not found"; 
+                }
+                
             
             } catch(Exception ex) { 
                 response.Success = false; 
@@ -52,11 +73,11 @@ namespace dotnet_rpg.Services.CharacterService
             return response;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var response = new ServiceResponse<List<GetCharacterDto>>(); 
             var dbCharacters = await _context.Characters
-            .Where(c => c.User.Id == userId).ToListAsync(); 
+            .Where(c => c.User.Id == GetUserId()).ToListAsync(); 
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return response;
             
@@ -65,7 +86,8 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>(); 
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             //FirstOrDefault function = return the 1st elemt of a sequence or a default value if element isnt there 
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter); 
             return serviceResponse;
@@ -78,17 +100,25 @@ namespace dotnet_rpg.Services.CharacterService
             try{ 
             
                 Character character = await _context.Characters
+                    .Include(c => c.User)
                     .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                    
+                if(character.User.Id == GetUserId()){ 
+                    character.Name = updatedCharacter.Name; 
+                    character.HitPoints = updatedCharacter.HitPoints; 
+                    character.Strength = updatedCharacter.Strength; 
+                    character.Defense = updatedCharacter.Defense;
+                    character.Intelligence = updatedCharacter.Intelligence; 
+                    character.Class = updatedCharacter.Class; 
+                    
+                    await _context.SaveChangesAsync();
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else { 
+                    response.Success = false; 
+                    response.Message = "Character Not Found";
+                }
                 
-                character.Name = updatedCharacter.Name; 
-                character.HitPoints = updatedCharacter.HitPoints; 
-                character.Strength = updatedCharacter.Strength; 
-                character.Defense = updatedCharacter.Defense;
-                character.Intelligence = updatedCharacter.Intelligence; 
-                character.Class = updatedCharacter.Class; 
-                
-                await _context.SaveChangesAsync();
-                response.Data = _mapper.Map<GetCharacterDto>(character);
             
             } catch(Exception ex) { 
                 response.Success = false;
